refactor: split data classes out of GuardianViewModel into AppState.kt
  Move DayTime, TimeSlot, Mode, Schedule, NfcTag, BlockMode, AppState,
  PendingUnlock, and ActivationResult from the top of GuardianViewModel.kt
  into a new AppState.kt. Pure Kotlin, only kotlinx.serialization
  annotations imported.
