+++ app/src/main/java/com/andebugulin/nfcguard/GuardianViewModel.kt
@@ -7,107 +7,8 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
-import kotlinx.serialization.Serializable
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
-
-@Serializable
-data class DayTime(
-    val day: Int,
-    val startHour: Int,
-    val startMinute: Int,
-    val endHour: Int,
-    val endMinute: Int
-)
-
-@Serializable
-data class TimeSlot(
-    val dayTimes: List<DayTime>
-) {
-    val days: List<Int> get() = dayTimes.map { it.day }
-    val startHour: Int get() = dayTimes.firstOrNull()?.startHour ?: 9
-    val startMinute: Int get() = dayTimes.firstOrNull()?.startMinute ?: 0
-    val endHour: Int get() = dayTimes.firstOrNull()?.endHour ?: 23
-    val endMinute: Int get() = dayTimes.firstOrNull()?.endMinute ?: 59
-    fun getTimeForDay(day: Int): DayTime? = dayTimes.find { it.day == day }
-}
-
-@Serializable
-data class Mode(
-    val id: String,
-    val name: String,
-    val blockedApps: List<String>,
-    val blockMode: BlockMode = BlockMode.BLOCK_SELECTED,
-    @Deprecated("Use nfcTagIds instead. Kept for migration from older versions.")
-    val nfcTagId: String? = null,
-    val nfcTagIds: List<String> = emptyList(),
-    val tagUnlockLimits: Map<String, Long?> = emptyMap() // tagId -> max unlock duration in minutes (null = permanent). Key "ANY" for any other tag.
-) {
-    /** Resolved tag list: migrates legacy single-tag field automatically. */
-    val effectiveNfcTagIds: List<String>
-        get() = if (nfcTagIds.isNotEmpty()) nfcTagIds
-        else if (@Suppress("DEPRECATION") nfcTagId != null) listOf(@Suppress("DEPRECATION") nfcTagId!!)
-        else emptyList()
-
-    /** Returns the limit for a specific tag, prioritizing exact match over wildcard. */
-    fun getLimitForTag(tagId: String): Long? {
-        if (tagUnlockLimits.containsKey(tagId)) return tagUnlockLimits[tagId]
-        if (tagUnlockLimits.containsKey("ANY")) return tagUnlockLimits["ANY"]
-        return null
-    }
-}
-
-@Serializable
-data class Schedule(
-    val id: String,
-    val name: String,
-    val timeSlot: TimeSlot,
-    val linkedModeIds: List<String>,
-    val hasEndTime: Boolean = false
-)
-
-@Serializable
-data class NfcTag(
-    val id: String,
-    val name: String,
-    val linkedModeIds: List<String> = emptyList()
-)
-
-@Serializable
-enum class BlockMode {
-    BLOCK_SELECTED,
-    ALLOW_SELECTED
-}
-
-@Serializable
-data class AppState(
-    val modes: List<Mode> = emptyList(),
-    val schedules: List<Schedule> = emptyList(),
-    val nfcTags: List<NfcTag> = emptyList(),
-    val activeModes: Set<String> = emptySet(),
-    val activeSchedules: Set<String> = emptySet(), // Schedules that activated their modes
-    val deactivatedSchedules: Set<String> = emptySet(), // Schedules manually deactivated by user
-    val manuallyActivatedModes: Set<String> = emptySet(), // Modes activated by user tap (not by schedule)
-    val timedModeDeactivations: Map<String, Long> = emptyMap(), // modeId -> epoch millis when it should auto-deactivate
-    val timedModeReactivations: Map<String, Long> = emptyMap(), // modeId -> epoch millis when it should auto-reactivate after NFC unlock
-    val pausedModeRemainingMs: Map<String, Long> = emptyMap() // modeId -> remaining deactivation ms saved when mode was paused
-)
-
-/** Pending NFC unlock awaiting user duration choice (not persisted) */
-data class PendingUnlock(
-    val modeIds: Set<String>,
-    val schedulesToDeactivate: Set<String>,
-    val tagId: String? = null, // Store which tag was scanned to apply its specific limit
-    val maxLimitMinutes: Long? = null, // The most restrictive limit among all modes being unlocked
-    val modeLimits: Map<String, Long?> = emptyMap() // per-mode limit: modeId -> limit (null = permanent)
-)
-
-/** Result of attempting to activate a mode */
-enum class ActivationResult {
-    SUCCESS,
-    BLOCK_MODE_CONFLICT,
-    MODE_NOT_FOUND
-}
 
 class GuardianViewModel(application: Application) : AndroidViewModel(application) {
